Move Page related functionality to @markbind/core
