+++ package.json
@@ -61,14 +61,12 @@
     "cheerio": "^0.22.0",
     "chokidar": "^3.3.0",
     "commander": "^3.0.2",
-    "fastmatter": "^2.1.1",
     "figlet": "^1.2.4",
     "find-up": "^4.1.0",
     "fs-extra-promise": "^1.0.1",
     "gh-pages": "^2.1.1",
     "htmlparser2": "^3.10.1",
     "ignore": "^5.1.4",
-    "js-beautify": "^1.10.3",
     "live-server": "^1.2.1",
     "lodash": "^4.17.15",
     "progress": "^2.0.3",
