+++ packages/core/package.json
@@ -27,6 +27,7 @@
     "fs-extra-promise": "^1.0.1",
     "highlight.js": "^9.14.2",
     "htmlparser2": "^3.10.1",
+    "js-beautify": "^1.10.3",
     "lodash": "^4.17.15",
     "markdown-it": "^11.0.0",
     "markdown-it-attrs": "^3.0.3",
