+++ packages/core/src/Page/__mocks__/index.js
@@ -1,6 +1,6 @@
 // mock generate
 
-const Page = jest.genMockFromModule('../Page');
+const Page = jest.genMockFromModule('../index');
 
 Page.prototype.generate = () => Promise.resolve('');
